requirements.txt
# Core ML & FL Framework
torch>=1.13.0
flwr[simulation]>=1.0.0  # Include simulation dependencies (ray)

# Data Processing
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.14.0

# Experiment Tracking (Optional)
wandb>=0.15.0
mlflow>=2.0.0

# Configuration
pyyaml>=6.0

# Utilities
tqdm>=4.64.0
scipy>=1.10.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.20.0

